Eliminated outdated code, added code for effect size, overlapping coefficient, and genus analysis
Eliminated code that created an now unused dataframe. Added code for the
calculation of effect size and overlapping coefficient of the leaf
length distributions. Added code to analyse the data at a genus level to
confirm that pines behave differently from other species. Reordered the
code to match the summary.