Fix #33 Add conf intervals to decomp-sr plot
I cleaned up the code in decomp-flam quite a bit to make it simpler.
This script produces "flamdecomp", a df that stores observed spread
rates for year 0 and predicted for year 1. Both with upper+ lower cf.

I'm using type="confidence" rather than type="prediction". We may
need to change that.