Added new files for the new volatile analysis
Read-vol2.R is meant to be an intermediate step. For now it holds code
to create summary dataframes, do the stats, and create plots.
The two csv files contain data used for the non-additivity question.
vol_pred_mix.csv has the calculated volatile content for the mixtures in
the master's project. fla_obs-pred.csv contains the difference between
observed and predicted values of flamambility for all the mixtures in
the masters project.