Re-add burn_moisture_trials_new
This is the file that Rita used for her thesis figs.  Its relationship
to burn_mositure_trials.csv is tricky to discern exactly.  It lacks the
start time and date information, so it is not possible to line these
observations up with the HOBO data.  But it also includes 12
observations not in burn_moisture_trials.csv.