Solve units issue with spread rate
It turns out that one of the metadata files stored with data had
incorrect information on units. Also, the description of some variables
was incorrect. A root issue is that the actual raw data is all missing,
the data as committed by Rita already includes calculations such as
spread rate and is missing the raw flaming start and end times, etc. I
must assume those simple calculations are correct, I am sure they are,
but it would be best to have the raw data with stopwatch times.