Correct metadata description moisture trials
This is still messy (same data in two files, one with _new appended).
But I am trying to leave all of Rita's data as is. We are missing the
actual raw data, but this version clearly worked up in ms excel first,
is the best we can do. The excel calculations were minor and I ahve no
worries (duration of flaming combustion, etc from the raw stopwatch
data).